+++ CHANGELOG.md
@@ -7,6 +7,8 @@ CHANGELOG
 ### Latest changes
 
 * VFS-10333 Unified naming of test suites
+* VFS-9688 Removed arrow placement from one-tooltip
+* VFS-10118 Added support for lambda configuration parameters
 * VFS-9305 Refactored one-tab-bar to not use jQuery
 * VFS-10103 Changed showing clipboard in popovers info
 * VFS-10367 Updated gui common